Renamed/moved accessors; fixed paths
