+++ src/Tasks/TrackerUpdates/Commissions/CommissionsTrackerTask.php
@@ -9,7 +9,7 @@ use App\Entity\ArtisanUrl;
 use App\Repository\ArtisanRepository;
 use App\Service\WebpageSnapshotManager;
 use App\Tasks\TrackerUpdates\TrackerTaskInterface;
-use App\Utils\Accessors\Commission;
+use App\Utils\Artisan\Fields\CommissionAccessor;
 use App\Utils\Artisan\SmartAccessDecorator as Artisan;
 use App\Utils\Data\ArtisanChanges;
 use App\Utils\StringList;
@@ -177,7 +177,7 @@ class CommissionsTrackerTask implements TrackerTaskInterface
 
             $newValue = StringList::pack(array_map(fn (OfferStatus $item): string => ucfirst(strtolower($item->getOffer())), $offersMatchingStatus));
 
-            Commission::set($result->getChanged(), $status, $newValue);
+            CommissionAccessor::set($result->getChanged(), $status, $newValue);
         }
 
         return $result;
