+++ tests/DataDefinitions/FieldsTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Utils\Artisan;
+namespace App\Tests\DataDefinitions;
 
 use App\DataDefinitions\Fields;
 use PHPUnit\Framework\TestCase;
@@ -12,7 +12,6 @@ use TRegx\CleanRegex\Match\Details\Detail;
 class FieldsTest extends TestCase
 {
     private const FIELD_NAME_REGEXP = 'const +(?<name>[A-Z_0-9]+) +=';
-    public const SRC_DIR_PATH = __DIR__.'/../../../src';
 
     /**
      * Purpose of this test is to make sure no field definition is accidentally left out.
@@ -22,7 +21,8 @@ class FieldsTest extends TestCase
     public function testIfTheDefinitionsAreComplete(): void
     {
         $classPath = str_replace(['App\\', '\\'], '/', Fields::class).'.php';
-        $modelSource = file_get_contents(self::SRC_DIR_PATH.'/'.$classPath);
+        $pathToRoot = str_repeat('../', substr_count(self::class, '\\') - 1);
+        $modelSource = file_get_contents(__DIR__."/$pathToRoot/src/$classPath");
 
         $fields = Fields::getAll();
